POC rusty-panel
rusty-panel is a small GTK panel written in Rust that shows a single
line of output from an arbitrary command.
I intended to use this as an replacement for the xfce4-panel.
Configuration of desktop is usually done through graphical menus often and
limited by the available plugins. Since this panel is showing the ouput
a simple shell script (or whatever command you configure) it is dead
easy to customize.